+++ lib/blocs/auth/registration/registration_bloc.dart
@@ -1,6 +1,6 @@
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
-import '../../repositories/auth_repository.dart';
+import '../../../repositories/auth_repository.dart';
 
 part 'registration_event.dart';
 part 'registration_state.dart';
