+++ lib/models/artikel_models.dart
@@ -1,5 +1,3 @@
-import 'dart:convert';
-
 import 'package:intl/intl.dart';
 
 class Article{
