+++ lib/models/catatan_dokter_model.dart
@@ -1,5 +1,3 @@
-import 'package:flutter/material.dart';
-
 class CatatanDokterModel{
   String gejala;
   String diagnosis;
