+++ lib/screens/halamanRegistrasi.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:rumah_sakit/blocs/registration/registration_bloc.dart';
+import 'package:rumah_sakit/blocs/auth/registration/registration_bloc.dart';
 import 'package:rumah_sakit/screens/halamanLogin.dart';
 import 'package:rumah_sakit/components/popupcustom.dart';
